[iOS] fix: 파일정리 및 코딩컨벤션 규칙에 따른 코드 수정
